Rename data_parser to yaml_parser for non-confusing names.
